ci: Enhance translation workflows by adding variables.json copy and updating file handling in filterUpdateFiles.js
